mark int type explicitly as int64 (#332)
* mark int type explicitly as int64

fixes cube2222#330


* use 32bit truncation for parquet

---------

+++ aggregates/count.go
@@ -15,7 +15,7 @@ var CountOverloads = []physical.AggregateDescriptor{
 }
 
 type Count struct {
-	count int
+	count int64
 }
 
 func NewCountPrototype() func() nodes.Aggregate {
